requirements.txt
sqlalchemy==1.3.13
pymysql==0.9.3
greenlet==0.4.15